Gulp factoring - Moved tasks into separate files - Moved all configuration into gulp/config.js
