+++ src/Umbraco.Web.UI.Client/gulp/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var fs = require('fs');
+
+var onlyScripts = require('./util/scriptFilter');
+var tasks = fs.readdirSync('./gulp/tasks/').filter(onlyScripts);
+
+tasks.forEach(function(task) {
+	require('./tasks/' + task);
+});
