+++ src/Umbraco.Web.UI.Client/gulp/tasks/test.js
@@ -1,0 +1,26 @@
+'use strict';
+
+var config = require('../config');
+var gulp = require('gulp');
+var karmaServer = require('karma').Server;
+
+/**************************
+ * Build tests
+ **************************/
+
+ // Karma test
+gulp.task('test:unit', function() {
+    new karmaServer({
+        configFile: __dirname + "/test/config/karma.conf.js",
+        keepalive: true
+    })
+    .start();
+});
+
+gulp.task('test:e2e', function() {
+    new karmaServer({
+        configFile: __dirname + "/test/config/e2e.js",
+        keepalive: true
+    })
+    .start();
+});
