+++ src/Umbraco.Web.UI.Client/gulp/util/handleErrors.js
@@ -1,0 +1,18 @@
+'use strict';
+
+var notify = require('gulp-notify');
+
+module.exports = function(error) {
+
+	var args = Array.prototype.slice.call(arguments);
+
+	// Send error to notification center with gulp-notify
+	notify.onError({
+		title: 'Compile Error',
+		message: '<%= error.message %>'
+	}).apply(this, args);
+
+	// Keep gulp from hanging on this task
+	this.emit('end');
+
+};
