+++ src/Umbraco.Web.UI.Client/gulp/util/scriptFilter.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var path = require('path');
+
+// Filters out non .js files. Prevents
+// accidental inclusion of possible hidden files
+module.exports = function(name) {
+	return /(\.(js)$)/i.test(path.extname(name));
+};
